// src/types/collections/Director.type.ts
// This interface defines the structure of a Director object
// It includes the director's ID, full name, associated movie and series IDs, birthplace, birthday, and notes
// This should be used for all operations except for the initial entry of a director in the form
export interface Director {
  id: string; // Unique string, 20 alphanumeric characters
  fullName: string; // Full name of the director
  movieIds: string[]; // List of movie IDs the director is associated with
  seriesIds: string[]; // List of series IDs the director is associated with
  birthplace: string; // Birthplace of the director
  birthday: string; // Birthday of the director
  notes: string; // Additional notes about the director
}

// This interface is used for the initial entry of a director in the form
// It is similar to the Director interface but does not include the id field
// as it will be generated by the Firestore service when the document is created
export interface DirectorInitialEntry {
    fullName: string; // Full name of the director
    movieIds: string[]; // List of movie IDs the director is associated with
    seriesIds: string[]; // List of series IDs the director is associated with
    birthplace: string; // Birthplace of the director
    birthday: string; // Birthday of the director
    notes: string; // Additional notes about the director
}